+++ backend/cmd/user.go
@@ -37,9 +37,6 @@ func validateUserInfo() {
 		if updateLoginType(user) {
 			updateUser = true
 		}
-		if updateUserDefaults(user) {
-			updateUser = true
-		}
 		adminUser := settings.Config.Auth.AdminUsername
 		adminPass := settings.Config.Auth.AdminPassword
 		passwordEnabled := settings.Config.Auth.Methods.PasswordAuth.Enabled
@@ -64,14 +61,6 @@ func validateUserInfo() {
 			}
 		}
 	}
-}
-
-func updateUserDefaults(user *users.User) bool {
-	if user.DisableOfficePreviewExt != "" && user.DisablePreviewExt == "" {
-		user.DisablePreviewExt = user.DisableOfficePreviewExt
-		return true
-	}
-	return false
 }
 
 func updateUserScopes(user *users.User) bool {